Make reionization modifiers use FOF group mass

+++ meraxes/src/core/tocf_reionization.c
@@ -236,7 +236,7 @@ void free_reionization_grids(run_globals_t *run_globals)
 }
 
 
-int find_cell(double pos, double box_size)
+int find_cell(float pos, double box_size)
 {
   int HII_dim = tocf_params.HII_dim;
