+++ meraxes/src/physics/mergers.c
@@ -50,10 +50,10 @@ double calculate_merging_time(run_globals_t *run_globals, galaxy_t *sat, int sna
   // timescale ala Guo+ 2010
   sat_mass = sat->Mvir + sat->ColdGas + sat->StellarMass;
 
-  sat_rad = sqrt(
-    pow(parent->Pos[0] - sat->Pos[0], 2.0) +
-    pow(parent->Pos[1] - sat->Pos[1], 2.0) +
-    pow(parent->Pos[2] - sat->Pos[2], 2.0));
+  sat_rad = (double)sqrt(
+    powf(parent->Pos[0] - sat->Pos[0], 2.0) +
+    powf(parent->Pos[1] - sat->Pos[1], 2.0) +
+    powf(parent->Pos[2] - sat->Pos[2], 2.0));
 
   // convert to physical length
   // Note that we want to use the redshift corresponding to the previous
